+++ src/components/PageNotFound/PageNotFound.css
@@ -15,7 +15,7 @@
 .ErrorPageText{
     display: grid;
     grid-template-columns: 1fr;
-    gap: 1rem;
+    gap: .5rem;
     text-align: center;
 }
