+++ src/router.js
@@ -40,7 +40,7 @@ import ProjectMetricsPage from './components/ProjectMetricsPage';
 import AppLogsPage from './components/AppLogsPage';
 import AppNetworkPage from './components/AppNetworkPage';
 import TeamPage from './components/TeamPage';
-import PageNotFound from './components/404page';
+import PageNotFound from './components/PageNotFound';
 
 // Protected route should have token. If not, login.
 const ProtectedRoute = ({ isAllowed, ...props }) => (
